requirements.txt
# Core
fastapi==0.109.1
uvicorn==0.27.0

# Data
tweepy==4.14.0
python-dotenv==1.0.0

# NLP
transformers==4.37.2
torch==2.1.2  # CPU

# Dashboard
streamlit==1.29.0
plotly==5.18.0

# mongoose 
mongoose==8.14.0

# MongoDB
pymongo==4.5.0

# celery redis 
celery[redis]==6.2.2
